+++ Runtime/Code/ScriptableObjects/ExtendedEliteDef.cs
@@ -11,6 +11,15 @@ namespace MSU
     [CreateAssetMenu(fileName = "New ExtendedEliteDef", menuName = "MSU/ExtendedEliteDef")]
     public class ExtendedEliteDef : EliteDef
     {
+        [Tooltip("An Enum Flags for adding Elites to the base game's tiers.\n" +
+            "Tier1: The Tier that includes Blazing, Glacial, Overloading and Mending\n" +
+            "Tier1Honor: Same as Tier 1, but weaker, used for the Artifact of Honor." +
+            "Tier1_5: Same as Tier 1, but with the Gilded Elite included.\n" +
+            "Tier1_5Honor: Same as Tier 1_5, but weaker, used for the Artifact of Honor.\n" +
+            "Tier2: The Tier that includes Malachite, Celestine, and Bead.\n" +
+            "Lunar: The Tier that has the Perfected Elites.\n" +
+            "GlobalTier1: Shorthand for Tier1 and Tier1_5.\n" +
+            "GlobalTier1Honor: Shorthand for Tier1Honor and Tier1_5Honor.")]
         public VanillaTierFlags eliteTierFlags;
 
         [Tooltip("A texture ramp for this elite, which will be applied to its model.")]
@@ -49,15 +58,53 @@ namespace MSU
         [Flags]
         public enum VanillaTierFlags : uint
         {
+            /// <summary>
+            /// No tier.
+            /// </summary>
             None = 0u,
+
+            /// <summary>
+            /// The tier that includes Blazing, Glacial, Overloading and Mending elites.
+            /// </summary>
             Tier1 = 1u,
+
+            /// <summary>
+            /// <inheritdoc cref="Tier1"/>
+            /// <br></br>
+            /// Weaker variants, used for the Artifact of Honor
+            /// </summary>
             Tier1Honor = 2u,
+
+            /// <summary>
+            /// The tier that includes Blazing, Glacial, Overloading, Mending and Gilded elites.
+            /// </summary>
             Tier1_5 = 4u,
+
+            /// <summary>
+            /// <inheritdoc cref="Tier1_5"/>
+            /// <br></br>
+            /// Weaker variants, used for the Artifact of Honor
+            /// </summary>
             Tier1_5Honor = 8u,
+
+            /// <summary>
+            /// The tier that includes Malachite, Celestine and Bead elites.
+            /// </summary>
             Tier2 = 16u,
+
+            /// <summary>
+            /// The tier that includes Perfected elites.
+            /// </summary>
             Lunar = 32u,
 
+            /// <summary>
+            /// Shorthand for Tier1 | Tier_5. Used if you want your elite to spawn on both tiers
+            /// </summary>
             GlobalTier1 = Tier1 | Tier1_5,
+
+            /// <summary>
+            /// Shorthand for Tier1Honor | Tier1_5Honor. Used if you want your elite to spawn on both tiers
+            /// </summary>
             GlobalTier1Honor = Tier1Honor | Tier1_5Honor
         }